requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.41.0
line-bot-sdk==3.22.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
python-dotenv==1.2.1
